Enum foundationdb::directory::DirectoryError

pub enum DirectoryError {

    CannotModifyRootDirectory,
    DirectoryPrefixInUse,
    DirectoryDoesNotExists,
    NoPathProvided,
    DirAlreadyExists,
    PathDoesNotExists,
    ParentDirDoesNotExists,
    IncompatibleLayer,
    BadDestinationDirectory,
    Version(String),
    PrefixNotAllowed,
    CannotPrefixInPartition,
    CannotMoveRootDirectory,
    CannotMoveBetweenPartition,
    CannotMoveBetweenSubdirectory,
    PrefixNotEmpty,
    CannotCreateSubpath,
    CannotOpenDirectoryPartition,
    CannotGetKeyDirectoryPartition,
    CannotPackDirectoryPartition,
    CannotUnpackDirectoryPartition,
    CannotRangeDirectoryPartition,
    IoError(Error),
    FdbError(FdbError),
    HcaError(HcaError),
    PackError(PackError),
    Other(String),

}

The enumeration holding all possible errors from a Directory.

Variants

CannotModifyRootDirectory

cannot modify the root directory

DirectoryPrefixInUse

prefix is already used

DirectoryDoesNotExists

Directory does not exists

NoPathProvided

missing path.

DirAlreadyExists

tried to create an already existing path.

PathDoesNotExists

missing directory.

ParentDirDoesNotExists

Parent does not exists

IncompatibleLayer

the layer is incompatible.

BadDestinationDirectory

the destination directory cannot be a subdirectory of the source directory.

Version(String)

Bad directory version.

PrefixNotAllowed

cannot specify a prefix unless manual prefixes are enabled

CannotPrefixInPartition

cannot specify a prefix in a partition.

CannotMoveRootDirectory

the root directory cannot be moved

CannotMoveBetweenPartition

CannotMoveBetweenSubdirectory

the destination directory cannot be a subdirectory of the source directory

PrefixNotEmpty

Prefix is not empty

CannotCreateSubpath

Thrown when the subpath cannot be computed due to length errors

CannotOpenDirectoryPartition

cannot open subspace in the root of a directory partition

CannotGetKeyDirectoryPartition

cannot get key for the root of a directory partition

CannotPackDirectoryPartition

cannot pack for the root of a directory partition

CannotUnpackDirectoryPartition

cannot unpack keys using the root of a directory partition

CannotRangeDirectoryPartition

cannot get range for the root of a directory partition

IoError(Error)

FdbError(FdbError)

HcaError(HcaError)

PackError(PackError)

Other(String)