Struct foundationdb::directory::DirectoryPartition

pub struct DirectoryPartition { /* private fields */ }

A DirectoryPartition is a DirectorySubspace whose prefix is preprended to all of its descendant directoriesโs prefixes. It cannot be used as a Subspace. Instead, you must create at least one subdirectory to store content.

Implementations

impl DirectoryPartition

pub fn get_path(&self) -> &[String]

pub fn get_layer(&self) -> &[u8]