Struct foundationdb::directory::DirectorySubspace

pub struct DirectorySubspace { /* fields omitted */ }

A DirectorySubspace represents the contents of a directory, but it also remembers the path with which it was opened and offers convenience methods to operate on the directory at that path. An instance of DirectorySubspace can be used for all the usual subspace operations. It can also be used to operate on the directory with which it was opened.

Implementations

impl DirectorySubspace

pub fn new(
path: &[String],
prefix: Vec<u8>,
directory_layer: &DirectoryLayer,
layer: Vec<u8>
) -> Self

impl DirectorySubspace

pub fn subspace<T: TuplePack>(&self, t: &T) -> Subspace

pub fn bytes(&self) -> &[u8]

pub fn pack<T: TuplePack>(&self, t: &T) -> Vec<u8>

pub fn unpack<'de, T: TupleUnpack<'de>>(&self, key: &'de [u8]) -> PackResult<T>

pub fn range(&self) -> (Vec<u8>, Vec<u8>)

pub fn get_path(&self) -> &[String]

pub fn set_path(&mut self, path: Vec<String>)

pub fn get_layer(&self) -> &[u8]

pub fn is_start_of(&self, key: &[u8]) -> bool