Struct foundationdb::future::FdbAddress

#[repr(packed(1))]
pub struct FdbAddress { /* private fields */ }

An address owned by a foundationDB future

Because the data it represent is owned by the future in FdbAddresses, you can never own a FdbAddress directly, you can only have references to it. This way, you can never obtain a lifetime greater than the lifetime of the slice that gave you access to it.