Module foundationdb::mapped_key_values

Definitions of MappedKeyValues, used in api version 710 and more.

GetMappedRange is an experimental feature introduced in FDB 7.1. It is intended to improve the client throughput and reduce latency for a commonly used traffic pattern. An experiment with Record Layer shows that this optimization can get 4x client throughput on a certain workload.

More info can be found in the relevant documentation.

Structs

FdbMappedKeyValue

A KeyValue produced by a mapped operation, ownder by a Foundation Future.

FdbMappedValue

An FdbMappedValue that you can own.

FdbMappedValuesIter

An iterator of mapped keyvalues owned by a foundationDB future

MappedKeyValues

An slice of mapped keyvalues owned by a foundationDB future produced by the get_mapped method.