Struct foundationdb::RangeOption

pub struct RangeOption<'a> {
    pub begin: KeySelector<'a>,
    pub end: KeySelector<'a>,
    pub limit: Option<usize>,
    pub target_bytes: usize,
    pub mode: StreamingMode,
    pub reverse: bool,
    /* private fields */
}

RangeOption represents a query parameters for range scan query.

You can construct RangeOption easily:

use foundationdb::RangeOption;

let opt = RangeOption::from((b"begin".as_ref(), b"end".as_ref()));
let opt: RangeOption = (b"begin".as_ref()..b"end".as_ref()).into();
let opt = RangeOption {
    limit: Some(10),
    ..RangeOption::from((b"begin".as_ref(), b"end".as_ref()))
};

Fields
begin: KeySelector<'a>

The beginning of the range.

end: KeySelector<'a>

The end of the range.

limit: Option<usize>

If non-zero, indicates the maximum number of key-value pairs to return.

target_bytes: usize

If non-zero, indicates a (soft) cap on the combined number of bytes of keys and values to return for each item.

mode: StreamingMode

One of the options::StreamingMode values indicating how the caller would like the data in the range returned.

reverse: bool

If true, key-value pairs will be returned in reverse lexicographical order beginning at the end of the range.

Implementations

impl<'a> RangeOption<'a>

pub fn rev(self) -> Self

Reverses the range direction.

pub fn next_range(self, kvs: &FdbValues) -> Option<Self>