Module foundationdb::api

Configuration of foundationDB API and Network

Provides a safe wrapper around the following C API calls:

API versioning

Network

Structs

FdbApiBuilder
A Builder with which different versions of the Fdb C API can be initialized

NetworkBuilder
A Builder with which the foundationDB network event loop can be created

NetworkRunner
A foundationDB network event loop runner

NetworkStop
Allow to stop the associated and running NetworkRunner.

NetworkWait
A condition object that can wait for the associated NetworkRunner to actually run.

Functions

get_max_api_version
Returns the max api version of the underlying Fdb C API Client