Struct foundationdb::directory::DirectoryLayer

pub struct DirectoryLayer { /* private fields */ }

A DirectoryLayer defines a new root directory. The node subspace and content subspace control where the directory metadata and contents, respectively, are stored. The default root directory has a node subspace with raw prefix \xFE and a content subspace with no prefix.

Implementations

impl DirectoryLayer

pub fn new( node_subspace: Subspace, content_subspace: Subspace, allow_manual_prefixes: bool ) -> Self

pub fn get_path(&self) -> &[String]