Function foundationdb::boot

pub fn boot() -> FdbResult<NetworkAutoStop>

Initialize the FoundationDB Client API, this can only be called once per process.

Returns

Network which must be run before the Client is ready. Network::run will not return until the network is stopped with the associated Network::stop and should be run in a separate thread.

Examples

let network = foundationdb::boot().expect("failed to initialize Fdb");

// do some interesting things with the API...

drop(network)