Enum foundationdb::options::TransactionOption

#[non_exhaustive]
pub enum TransactionOption {

    CausalWriteRisky,
    CausalReadRisky,
    CausalReadDisable,
    IncludePortInAddress,
    NextWriteNoWriteConflictRange,
    ReadYourWritesDisable,
    ReadAheadDisable,
    ReadServerSideCacheEnable,
    ReadServerSideCacheDisable,
    ReadPriorityNormal,
    ReadPriorityLow,
    ReadPriorityHigh,
    DurabilityDatacenter,
    DurabilityRisky,
    DurabilityDevNullIsWebScale,
    PrioritySystemImmediate,
    PriorityBatch,
    InitializeNewDatabase,
    AccessSystemKeys,
    ReadSystemKeys,
    RawAccess,
    BypassStorageQuota,
    DebugRetryLogging(String),
    TransactionLoggingEnable(String),
    DebugTransactionIdentifier(String),
    LogTransaction,
    TransactionLoggingMaxFieldLength(i32),
    ServerRequestTracing,
    Timeout(i32),
    RetryLimit(i32),
    MaxRetryDelay(i32),
    SizeLimit(i32),
    AutomaticIdempotency,
    SnapshotRywEnable,
    SnapshotRywDisable,
    LockAware,
    UsedDuringCommitProtectionDisable,
    ReadLockAware,
    UseProvisionalProxies,
    ReportConflictingKeys,
    SpecialKeySpaceRelaxed,
    SpecialKeySpaceEnableWrites,
    Tag(String),
    AutoThrottleTag(String),
    SpanParent(Vec<u8>),
    ExpensiveClearCostEstimationEnable,
    BypassUnreadable,
    UseGrvCache,
    AuthorizationToken(String),

}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

CausalWriteRisky

The transaction, if not self-conflicting, may be committed a second time after commit succeeds, in the event of a fault

CausalReadRisky

The read version will be committed, and usually will be the latest committed, but might not be the latest committed in the event of a simultaneous fault and misbehaving clock.

CausalReadDisable

IncludePortInAddress

Addresses returned by get_addresses_for_key include the port when enabled. As of api version 630, this option is enabled by default and setting this has no effect.

NextWriteNoWriteConflictRange

The next write performed on this transaction will not generate a write conflict range. As a result, other transactions which read the key(s) being modified by the next write will not conflict with this transaction. Care needs to be taken when using this option on a transaction that is shared between multiple threads. When setting this option, write conflict ranges will be disabled on the next write operation, regardless of what thread it is on.

ReadYourWritesDisable

Reads performed by a transaction will not see any prior mutations that occured in that transaction, instead seeing the value which was in the database at the transaction’s read version. This option may provide a small performance benefit for the client, but also disables a number of client-side optimizations which are beneficial for transactions which tend to read and write the same keys within a single transaction. It is an error to set this option after performing any reads or writes on the transaction.

ReadAheadDisable

Deprecated

ReadServerSideCacheEnable

Storage server should cache disk blocks needed for subsequent read requests in this transaction. This is the default behavior.

ReadServerSideCacheDisable

Storage server should not cache disk blocks needed for subsequent read requests in this transaction. This can be used to avoid cache pollution for reads not expected to be repeated.

ReadPriorityNormal

Use normal read priority for subsequent read requests in this transaction. This is the default read priority.

ReadPriorityLow

Use low read priority for subsequent read requests in this transaction.

ReadPriorityHigh

Use high read priority for subsequent read requests in this transaction.

DurabilityDatacenter

DurabilityRisky

DurabilityDevNullIsWebScale

Deprecated

PrioritySystemImmediate

Specifies that this transaction should be treated as highest priority and that lower priority transactions should block behind this one. Use is discouraged outside of low-level tools

PriorityBatch

Specifies that this transaction should be treated as low priority and that default priority transactions will be processed first. Batch priority transactions will also be throttled at load levels smaller than for other types of transactions and may be fully cut off in the event of machine failures. Useful for doing batch work simultaneously with latency-sensitive work

InitializeNewDatabase

This is a write-only transaction which sets the initial configuration. This option is designed for use by database system tools only.

AccessSystemKeys

Allows this transaction to read and modify system keys (those that start with the byte 0xFF). Implies raw_access.

ReadSystemKeys

Allows this transaction to read system keys (those that start with the byte 0xFF). Implies raw_access.

RawAccess

Allows this transaction to access the raw key-space when tenant mode is on.

BypassStorageQuota

Allows this transaction to bypass storage quota enforcement. Should only be used for transactions that directly or indirectly decrease the size of the tenant group’s data.

DebugRetryLogging(String)

Optional transaction name

TransactionLoggingEnable(String)

String identifier to be used in the logs when tracing this transaction. The identifier must not exceed 100 characters.

Deprecated

DebugTransactionIdentifier(String)

String identifier to be used when tracing or profiling this transaction. The identifier must not exceed 100 characters.

Sets a client provided identifier for the transaction that will be used in scenarios like tracing or profiling. Client trace logging or transaction profiling must be separately enabled.

LogTransaction

Enables tracing for this transaction and logs results to the client trace logs. The DEBUG_TRANSACTION_IDENTIFIER option must be set before using this option, and client trace logging must be enabled to get log output.

TransactionLoggingMaxFieldLength(i32)

Maximum length of escaped key and value fields.

Sets the maximum escaped length of key and value fields to be logged to the trace file via the LOG_TRANSACTION option, after which the field will be truncated. A negative value disables truncation.

ServerRequestTracing

Sets an identifier for server tracing of this transaction. When committed, this identifier triggers logging when each part of the transaction authority encounters it, which is helpful in diagnosing slowness in misbehaving clusters. The identifier is randomly generated. When there is also a debug_transaction_identifier, both IDs are logged together.

Timeout(i32)

value in milliseconds of timeout

Set a timeout in milliseconds which, when elapsed, will cause the transaction automatically to be cancelled. Valid parameter values are [0, INT_MAX]. If set to 0, will disable all timeouts. All pending and any future uses of the transaction will throw an exception. The transaction can be used again after it is reset. Prior to API version 610, like all other transaction options, the timeout must be reset after a call to onError. If the API version is 610 or greater, the timeout is not reset after an onError call. This allows the user to specify a longer timeout on specific transactions than the default timeout specified through the transaction_timeout database option without the shorter database timeout cancelling transactions that encounter a retryable error. Note that at all API versions, it is safe and legal to set the timeout each time the transaction begins, so most code written assuming the older behavior can be upgraded to the newer behavior without requiring any modification, and the caller is not required to implement special logic in retry loops to only conditionally set this option.

RetryLimit(i32)

number of times to retry

Set a maximum number of retries after which additional calls to onError will throw the most recently seen error code. Valid parameter values are [-1, INT_MAX]. If set to -1, will disable the retry limit. Prior to API version 610, like all other transaction options, the retry limit must be reset after a call to onError. If the API version is 610 or greater, the retry limit is not reset after an onError call. Note that at all API versions, it is safe and legal to set the retry limit each time the transaction begins, so most code written assuming the older behavior can be upgraded to the newer behavior without requiring any modification, and the caller is not required to implement special logic in retry loops to only conditionally set this option.

MaxRetryDelay(i32)

value in milliseconds of maximum delay

Set the maximum amount of backoff delay incurred in the call to onError if the error is retryable. Defaults to 1000 ms. Valid parameter values are [0, INT_MAX]. If the maximum retry delay is less than the current retry delay of the transaction, then the current retry delay will be clamped to the maximum retry delay. Prior to API version 610, like all other transaction options, the maximum retry delay must be reset after a call to onError. If the API version is 610 or greater, the retry limit is not reset after an onError call. Note that at all API versions, it is safe and legal to set the maximum retry delay each time the transaction begins, so most code written assuming the older behavior can be upgraded to the newer behavior without requiring any modification, and the caller is not required to implement special logic in retry loops to only conditionally set this option.

SizeLimit(i32)

value in bytes

Set the transaction size limit in bytes. The size is calculated by combining the sizes of all keys and values written or mutated, all key ranges cleared, and all read and write conflict ranges. (In other words, it includes the total size of all data included in the request to the cluster to commit the transaction.) Large transactions can cause performance problems on FoundationDB clusters, so setting this limit to a smaller value than the default can help prevent the client from accidentally degrading the cluster’s performance. This value must be at least 32 and cannot be set to higher than 10,000,000, the default transaction size limit.

AutomaticIdempotency

Automatically assign a random 16 byte idempotency id for this transaction. Prevents commits from failing with commit_unknown_result. WARNING: If you are also using the multiversion client or transaction timeouts, if either cluster_version_changed or transaction_timed_out was thrown during a commit, then that commit may have already succeeded or may succeed in the future. This feature is in development and not ready for general use.

SnapshotRywEnable

Snapshot read operations will see the results of writes done in the same transaction. This is the default behavior.

SnapshotRywDisable

Snapshot read operations will not see the results of writes done in the same transaction. This was the default behavior prior to API version 300.

LockAware

The transaction can read and write to locked databases, and is responsible for checking that it took the lock.

UsedDuringCommitProtectionDisable

By default, operations that are performed on a transaction while it is being committed will not only fail themselves, but they will attempt to fail other in-flight operations (such as the commit) as well. This behavior is intended to help developers discover situations where operations could be unintentionally executed after the transaction has been reset. Setting this option removes that protection, causing only the offending operation to fail.

ReadLockAware

The transaction can read from locked databases.

UseProvisionalProxies

This option should only be used by tools which change the database configuration.

ReportConflictingKeys

The transaction can retrieve keys that are conflicting with other transactions.

SpecialKeySpaceRelaxed

By default, the special key space will only allow users to read from exactly one module (a subspace in the special key space). Use this option to allow reading from zero or more modules. Users who set this option should be prepared for new modules, which may have different behaviors than the modules they’re currently reading. For example, a new module might block or return an error.

SpecialKeySpaceEnableWrites

By default, users are not allowed to write to special keys. Enable this option will implicitly enable all options required to achieve the configuration change.

Tag(String)

String identifier used to associated this transaction with a throttling group. Must not exceed 16 characters.

Adds a tag to the transaction that can be used to apply manual targeted throttling. At most 5 tags can be set on a transaction.

AutoThrottleTag(String)

String identifier used to associated this transaction with a throttling group. Must not exceed 16 characters.

Adds a tag to the transaction that can be used to apply manual or automatic targeted throttling. At most 5 tags can be set on a transaction.

SpanParent(Vec<u8>)

A byte string of length 16 used to associate the span of this transaction with a parent

Adds a parent to the Span of this transaction. Used for transaction tracing. A span can be identified with any 16 bytes

ExpensiveClearCostEstimationEnable

Asks storage servers for how many bytes a clear key range contains. Otherwise uses the location cache to roughly estimate this.

BypassUnreadable

Allows get operations to read from sections of keyspace that have become unreadable because of versionstamp operations. These reads will view versionstamp operations as if they were set operations that did not fill in the versionstamp.

UseGrvCache

Allows this transaction to use cached GRV from the database context. Defaults to off. Upon first usage, starts a background updater to periodically update the cache to avoid stale read versions. The disable_client_bypass option must also be set.

AuthorizationToken(String)

A JSON Web Token authorized to access data belonging to one or more tenants, indicated by ‘tenants’ claim of the token’s payload.

Attach given authorization token to the transaction such that subsequent tenant-aware requests are authorized

Implementations

impl TransactionOption

pub fn code(&self) -> FDBTransactionOption

pub unsafe fn apply(&self, target: *mut FDBTransaction) -> FdbResult<()>