Struct foundationdb::Transaction

pub struct Transaction { /* private fields */ }

In FoundationDB, a transaction is a mutable snapshot of a database.

All read and write operations on a transaction see and modify an otherwise-unchanging version of the database and only change the underlying database if and when the transaction is committed. Read operations do see the effects of previous write operations on the same transaction. Committing a transaction usually succeeds in the absence of conflicts.

Applications must provide error handling and an appropriate retry loop around the application code for a transaction. See the documentation for fdb_transaction_on_error().

Transactions group operations into a unit with the properties of atomicity, isolation, and durability. Transactions also provide the ability to maintain an application’s invariants or integrity constraints, supporting the property of consistency. Together these properties are known as ACID.

Transactions are also causally consistent: once a transaction has been successfully committed, all subsequently created transactions will see the modifications made by it.

Implementations

impl Transaction

pub fn set_option(&self, opt: TransactionOption) -> FdbResult<()>

Called to set an option on an FDBTransaction.

pub fn set_raw_option( &self, code: FDBTransactionOption, data: Option<Vec<u8>>, ) -> FdbResult<()>

Pass through an option given a code and raw data. Useful when creating a passthrough layer where the code and data will be provided as raw, in order to avoid deserializing to an option and serializing it back to code and data. In general, you should use set_option.

pub fn set(&self, key: &[u8], value: &[u8])

Modify the database snapshot represented by transaction to change the given key to have the given value.

If the given key was not previously present in the database it is inserted. The modification affects the actual database only if transaction is later committed with Transaction::commit.

Arguments

key - the name of the key to be inserted into the database.

value - the value to be inserted into the database

pub fn clear(&self, key: &[u8])

Modify the database snapshot represented by transaction to remove the given key from the database.

If the key was not previously present in the database, there is no effect. The modification affects the actual database only if transaction is later committed with Transaction::commit.

Arguments

key - the name of the key to be removed from the database.

pub fn get( &self, key: &[u8], snapshot: bool, ) -> impl Future<Output = FdbResult<Option<FdbSlice>>> + Send + Sync + Unpin

Reads a value from the database snapshot represented by transaction.

Returns an FDBFuture which will be set to the value of key in the database if there is any.

Arguments

key - the name of the key to be looked up in the database

snapshot - true if this is a snapshot read

pub fn atomic_op(&self, key: &[u8], param: &[u8], op_type: MutationType)

Modify the database snapshot represented by transaction to perform the operation indicated by operationType with operand param to the value stored by the given key.

An atomic operation is a single database command that carries out several logical steps: reading the value of a key, performing a transformation on that value, and writing the result. Different atomic operations perform different transformations. Like other database operations, an atomic operation is used within a transaction; however, its use within a transaction will not cause the transaction to conflict.

Atomic operations do not expose the current value of the key to the client but simply send the database the transformation to apply. In regard to conflict checking, an atomic operation is equivalent to a write without a read. It can only cause other transactions performing reads of the key to conflict.

By combining these logical steps into a single, read-free operation, FoundationDB can guarantee that the transaction will not conflict due to the operation. This makes atomic operations ideal for operating on keys that are frequently modified. A common example is the use of a key-value pair as a counter.

Warning

If a transaction uses both an atomic operation and a strictly serializable read on the same key, the benefits of using the atomic operation (for both conflict checking and performance) are lost.

pub fn get_key( &self, selector: &KeySelector<'_>, snapshot: bool, ) -> impl Future<Output = FdbResult<FdbSlice>> + Send + Sync + Unpin

Resolves a key selector against the keys in the database snapshot represented by transaction.

Returns an FDBFuture which will be set to the key in the database matching the key selector.

Arguments

selector: the key selector

snapshot: true if this is a snapshot read

pub fn get_ranges<'a>( &'a self, opt: RangeOption<'a>, snapshot: bool, ) -> impl Stream<Item = FdbResult<FdbValues>> + Send + Sync + Unpin + 'a

Reads all key-value pairs in the database snapshot represented by transaction (potentially limited by limit, target_bytes, or mode) which have a key lexicographically greater than or equal to the key resolved by the begin key selector and lexicographically less than the key resolved by the end key selector.

Returns a stream of KeyValue slices.

This method is a little more efficient than get_ranges_keyvalues but a little harder to use.

Arguments

opt: the range, limit, target_bytes and mode

snapshot: true if this is a snapshot read

pub fn get_ranges_keyvalues<'a>( &'a self, opt: RangeOption<'a>, snapshot: bool, ) -> impl Stream<Item = FdbResult<FdbValue>> + Unpin + 'a

Reads all key-value pairs in the database snapshot represented by transaction (potentially limited by limit, target_bytes, or mode) which have a key lexicographically greater than or equal to the key resolved by the begin key selector and lexicographically less than the key resolved by the end key selector.

Returns a stream of KeyValue.

Arguments

opt: the range, limit, target_bytes and mode

snapshot: true if this is a snapshot read

pub fn get_range( &self, opt: &RangeOption<'_>, iteration: usize, snapshot: bool, ) -> impl Future<Output = FdbResult<FdbValues>> + Send + Sync + Unpin

Reads all key-value pairs in the database snapshot represented by transaction (potentially limited by limit, target_bytes, or mode) which have a key lexicographically greater than or equal to the key resolved by the begin key selector and lexicographically less than the key resolved by the end key selector.

Arguments

opt: the range, limit, target_bytes and mode

iteration: If opt.mode is Iterator, this parameter should start at 1 and be incremented by 1 for each successive call while reading this range. In all other cases it is ignored.

snapshot: true if this is a snapshot read

pub fn get_mapped_range( &self, opt: &RangeOption<'_>, mapper: &[u8], iteration: usize, snapshot: bool, ) -> impl Future<Output = FdbResult<MappedKeyValues>> + Send + Sync + Unpin

Mapped Range is an experimental feature introduced in FDB 7.1. It is intended to improve the client throughput and reduce latency for querying data through a Subspace used as a “index”. In such a case, querying records by scanning an index in relational databases can be translated to a GetRange request on the index entries followed up by multiple GetValue requests for the record entries in FDB.

This method is allowing FoundationDB “follow up” a GetRange request with GetValue requests, this can happen in one request without additional back and forth. Considering the overhead of each request, this saves time and resources on serialization, deserialization, and network.

A mapped request will:

Do a range query (same as a Transaction.get_range request) and get the result. We call it the primary query.

For each key-value pair in the primary query result, translate it to a get_range query and get the result. We call them secondary queries.

Put all results in a nested structure and return them.

WARNING : This feature is considered experimental at this time. It is only allowed when using snapshot isolation AND disabling read-your-writes.

More info can be found in the relevant documentation.

This is the “raw” version, users are expected to use Transaction::get_mapped_ranges

pub fn get_mapped_ranges<'a>( &'a self, opt: RangeOption<'a>, mapper: &'a [u8], snapshot: bool, ) -> impl Stream<Item = FdbResult<MappedKeyValues>> + Send + Sync + Unpin + 'a

Mapped Range is an experimental feature introduced in FDB 7.1. It is intended to improve the client throughput and reduce latency for querying data through a Subspace used as a “index”. In such a case, querying records by scanning an index in relational databases can be translated to a GetRange request on the index entries followed up by multiple GetValue requests for the record entries in FDB.

This method is allowing FoundationDB “follow up” a GetRange request with GetValue requests, this can happen in one request without additional back and forth. Considering the overhead of each request, this saves time and resources on serialization, deserialization, and network.

A mapped request will:

Do a range query (same as a Transaction.get_range request) and get the result. We call it the primary query.

For each key-value pair in the primary query result, translate it to a get_range query and get the result. We call them secondary queries.

Put all results in a nested structure and return them.

WARNING : This feature is considered experimental at this time. It is only allowed when using snapshot isolation AND disabling read-your-writes.

More info can be found in the relevant documentation.

pub fn clear_range(&self, begin: &[u8], end: &[u8])

Modify the database snapshot represented by transaction to remove all keys (if any) which are lexicographically greater than or equal to the given begin key and lexicographically less than the given end_key.

The modification affects the actual database only if transaction is later committed with Transaction::commit.

pub fn get_estimated_range_size_bytes( &self, begin: &[u8], end: &[u8], ) -> impl Future<Output = FdbResult<i64>> + Send + Sync + Unpin

Get the estimated byte size of the key range based on the byte sample collected by FDB

pub fn commit( self, ) -> impl Future<Output = Result<TransactionCommitted, TransactionCommitError>> + Send + Sync + Unpin

Attempts to commit the sets and clears previously applied to the database snapshot represented by transaction to the actual database.

The commit may or may not succeed – in particular, if a conflicting transaction previously committed, then the commit must fail in order to preserve transactional isolation. If the commit does succeed, the transaction is durably committed to the database and all subsequently started transactions will observe its effects.

It is not necessary to commit a read-only transaction – you can simply drop it.

Callers will usually want to retry a transaction if the commit or a another method on the transaction returns a retryable error (see on_error and/or Database::transact).

As with other client/server databases, in some failure scenarios a client may be unable to determine whether a transaction succeeded. In these cases, Transaction::commit will return an error and is_maybe_committed() will returns true on that error. The on_error function treats this error as retryable, so retry loops that don’t check for is_maybe_committed() could execute the transaction twice. In these cases, you must consider the idempotence of the transaction. For more information, see Transactions with unknown results.

Normally, commit will wait for outstanding reads to return. However, if those reads were snapshot reads or the transaction option for disabling “read-your-writes” has been invoked, any outstanding reads will immediately return errors.

pub fn on_error( self, err: FdbError, ) -> impl Future<Output = FdbResult<Transaction>> + Send + Sync + Unpin

Implements the recommended retry and backoff behavior for a transaction. This function knows which of the error codes generated by other Transaction functions represent temporary error conditions and which represent application errors that should be handled by the application. It also implements an exponential backoff strategy to avoid swamping the database cluster with excessive retries when there is a high level of conflict between transactions.

It is not necessary to call reset() when handling an error with on_error() since the transaction has already been reset.

You should not call this method most of the times and use Database::transact which implements a retry loop strategy for you.

pub fn cancel(self) -> TransactionCancelled

Cancels the transaction. All pending or future uses of the transaction will return a transaction_cancelled error. The transaction can be used again after it is reset.

pub fn get_addresses_for_key( &self, key: &[u8], ) -> impl Future<Output = FdbResult<FdbAddresses>> + Send + Sync + Unpin

Returns a list of public network addresses as strings, one for each of the storage servers responsible for storing key_name and its associated value.

pub fn watch( &self, key: &[u8], ) -> impl Future<Output = FdbResult<()>> + Send + Sync + Unpin

A watch’s behavior is relative to the transaction that created it. A watch will report a change in relation to the key’s value as readable by that transaction. The initial value used for comparison is either that of the transaction’s read version or the value as modified by the transaction itself prior to the creation of the watch. If the value changes and then changes back to its initial value, the watch might not report the change.

Until the transaction that created it has been committed, a watch will not report changes made by other transactions. In contrast, a watch will immediately report changes made by the transaction itself. Watches cannot be created if the transaction has set the READ_YOUR_WRITES_DISABLE transaction option, and an attempt to do so will return an watches_disabled error.

If the transaction used to create a watch encounters an error during commit, then the watch will be set with that error. A transaction whose commit result is unknown will set all of its watches with the commit_unknown_result error. If an uncommitted transaction is reset or destroyed, then any watches it created will be set with the transaction_cancelled error.

Returns an future representing an empty value that will be set once the watch has detected a change to the value at the specified key.

By default, each database connection can have no more than 10,000 watches that have not yet reported a change. When this number is exceeded, an attempt to create a watch will return a too_many_watches error. This limit can be changed using the MAX_WATCHES database option. Because a watch outlives the transaction that creates it, any watch that is no longer needed should be cancelled by dropping its future.

pub fn get_approximate_size( &self, ) -> impl Future<Output = FdbResult<i64>> + Send + Sync + Unpin

Returns an FDBFuture which will be set to the approximate transaction size so far in the returned future, which is the summation of the estimated size of mutations, read conflict ranges, and write conflict ranges.

This can be called multiple times before the transaction is committed.

pub fn get_range_split_points( &self, begin: &[u8], end: &[u8], chunk_size: i64, ) -> impl Future<Output = FdbResult<FdbKeys>> + Send + Sync + Unpin

Gets a list of keys that can split the given range into (roughly) equally sized chunks based on chunk_size. Note: the returned split points contain the start key and end key of the given range.

pub fn get_versionstamp( &self, ) -> impl Future<Output = FdbResult<FdbSlice>> + Send + Sync + Unpin

Returns an FDBFuture which will be set to the versionstamp which was used by any versionstamp operations in this transaction.

The future will be ready only after the successful completion of a call to commit() on this Transaction. Read-only transactions do not modify the database when committed and will result in the future completing with an error. Keep in mind that a transaction which reads keys and then sets them to their current values may be optimized to a read-only transaction.

Most applications will not call this function.

pub fn get_read_version( &self, ) -> impl Future<Output = FdbResult<i64>> + Send + Sync + Unpin

The transaction obtains a snapshot read version automatically at the time of the first call to get_*() (including this one) and (unless causal consistency has been deliberately compromised by transaction options) is guaranteed to represent all transactions which were reported committed before that call.

pub fn set_read_version(&self, version: i64)

Sets the snapshot read version used by a transaction.

This is not needed in simple cases. If the given version is too old, subsequent reads will fail with error_code_past_version; if it is too new, subsequent reads may be delayed indefinitely and/or fail with error_code_future_version. If any of get_*() have been called on this transaction already, the result is undefined.

pub async fn get_metadata_version( &self, snapshot: bool, ) -> FdbResult<Option<i64>>

The metadata version key \xff/metadataVersion is a key intended to help layers deal with hot keys. The value of this key is sent to clients along with the read version from the proxy, so a client can read its value without communicating with a storage server. To retrieve the metadataVersion, you need to set TransactionOption::ReadSystemKeys

pub fn update_metadata_version(&self)

pub fn reset(&mut self)

Reset transaction to its initial state.

In order to protect against a race condition with cancel(), this call require a mutable access to the transaction.

This is similar to dropping the transaction and creating a new one.

It is not necessary to call reset() when handling an error with on_error() since the transaction has already been reset.

pub fn add_conflict_range( &self, begin: &[u8], end: &[u8], ty: ConflictRangeType, ) -> FdbResult<()>

Adds a conflict range to a transaction without performing the associated read or write.

Note

Most applications will use the serializable isolation that transactions provide by default and will not need to manipulate conflict ranges.

impl Transaction

pub fn clear_subspace_range(&self, subspace: &Subspace)