Module foundationdb::tuple

Implementation of the official tuple layer typecodes

The official specification can be found here.

Modules

hca

The directory layer offers subspace indirection, where logical application subspaces are mapped to short, auto-generated key prefixes. This prefix assignment is done by the High Contention Allocator, which allows many clients to allocate short directory prefixes efficiently.

Structs

Bytes

Represent a sequence of bytes (i.e. &u8)

Subspace

Represents a well-defined region of keyspace in a FoundationDB database

TupleDepth

Tracks the depth of a Tuple decoding chain

Uuid

A Universally Unique Identifier (UUID).

Versionstamp

Enums

Element

PackError

A packing/unpacking error

VersionstampOffset

Traits

TuplePack

A type that can be packed

TupleUnpack

A type that can be unpacked

Functions

pack

Pack value and returns the packed buffer

pack_into

Pack value into the given buffer

pack_into_with_versionstamp

Pack value into the given buffer

pack_with_versionstamp

Pack value and returns the packed buffer

unpack

Unpack input

Type Aliases

PackResult

Alias for Result<..., tuple::Error>